vm: fix constantinople tests + lint
